data/requirements.txt
arinc424 @ git+https://github.com/andrewda/arinc424@996025203aebf0a3794d2e337fe536ff4f4845b0
google-auth==2.16.3
numpy==2.0.1
pandas==2.2.2
pandas-gbq==0.23.1
requests==2.28.2
tqdm==4.64.1
beautifulsoup4==4.11.2
pytest==8.3.3
